yeat

+++ chapter8/src/exercise1/Faculty.java
@@ -2,7 +2,7 @@ package exercise1;
 
 public class Faculty extends UEmployee {
 private String deptName;
-	public Faculty(String fn, String ln, Double sal, String d) {
+	public Faculty(String fn, String ln, Integer sal, String d) {
 	
 		super(fn, ln,sal);
 		deptName = d;
