+++ chapter8/src/exercise1/UniTester.java
@@ -17,7 +17,7 @@ public class UniTester
     	String ln = input.next();
     	 System.out.println("Enter employee salary");
     	 
-    	 double sal = input.nextDouble();
+    	 int sal = input.nextInt();
 		
     	 System.out.println("Enter employee type: Faculty or Staff");
     	  String empT;
